lab6-rest-vs-grpc-Pradyoth-master/PradyothGRPC/requirements.txt
boto==2.44.0
chardet==2.3.0
crcmod==1.7
cryptography==1.7.1
Django==1.11.25
enum34==1.1.6
futures==3.3.0
google-compute-engine==20190916.0
grpc==0.3.post19
grpcio==1.24.3
grpcio-tools==1.24.3
idna==2.2
image==1.5.27
ipaddress==1.0.17
keyring==10.1
keyrings.alt==1.3
msgpack-python==0.5.6
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.1.9
pycrypto==2.6.1
pygobject==3.22.0
pytz==2019.3
pyxdg==0.25
requests==2.12.4
SecretStorage==2.3.1
six==1.12.0
urllib3==1.19.1